Offices and Users controllers

+++ Backend/HealthInsurance.Core/Services/UserService.cs
@@ -1,42 +1,47 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using AutoMapper;
+﻿using AutoMapper;
+using HealthInsurance.Core.Entities;
+using HealthInsurance.Core.Exceptions;
 using HealthInsurance.Core.Interfaces.Services;
 using HealthInsurance.Core.Interfaces.Specifications;
 using HealthInsurance.Core.Models;
-using HealthInsurance.Core.Entities;
-//using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace HealthInsurance.Core.Services
 {
-	public class UserService : IUserService
+    public class UserService : IUserService
 	{
-        //private ILogger _logger;
+        private ILogger _logger;
 		private IMapper _mapper;
 		private IUnitOfWork _unitOfWork;
 		private IRepository _repository;
 
-		public UserService(IMapper mapper, IUnitOfWork unitOfWork)
-		{
-            //_logger = logger;
-			_mapper = mapper;
+		public UserService(IUnitOfWork unitOfWork, IMapper mapper, ILogger<UserService> logger)
+		{   
 			_unitOfWork = unitOfWork;
 			_repository = unitOfWork.Repository;
-		}
+            _mapper = mapper;
+            _logger = logger;
+        }
 
 		public async Task<UserDto> GetById(int id)
 		{
 			var user = await _repository.GetById<User>(id);
-			var userDto = Mapper.Map<UserDto>(user);
-			return userDto;
+
+            if(user == null)
+            {
+                throw new NotFoundException($"User with Id = {id} is not found");
+            }
+
+			return Mapper.Map<UserDto>(user);
 		}
 
 		public async Task<IReadOnlyList<UserDto>> GetAll()
 		{
-            //_logger.LogWarning("Something is not gooood :( :(!");
 			var users = await _repository.GetAll<User>();
-			var usersDto = Mapper.Map<IReadOnlyList<UserDto>>(users);
-			return usersDto;
+
+			return Mapper.Map<IReadOnlyList<UserDto>>(users);
 		}
 
 		public async Task<UserDto> Add(UserForCreationDto user)
@@ -62,7 +67,13 @@ namespace HealthInsurance.Core.Services
 		public async Task<UserDto> Delete(int id)
 		{
 			var user = await _repository.GetById<User>(id);
-			_repository.Delete(user);
+
+            if (user == null)
+            {
+                throw new NotFoundException($"User with Id = {id} is not found");
+            }
+
+            _repository.Delete(user);
 
 			return _mapper.Map<UserDto>(user);
 		}
